Move schema to initialize back to individual migration
